Change TestMesh to AssimpMesh; no change in the output
