+++ game_client/AssimpModel.h
@@ -1,34 +1,37 @@
-#ifndef ASSIMP_MODEL_H
-#define ASSIMPMODELH
+////////////////////////////////////////
+// AssimpModel.h
+// 
+// Class that imports a model from an external file by Assimp
+////////////////////////////////////////
 
-#include <glm/glm.hpp>
-#include <glm/gtc/matrix_transform.hpp>
+#pragma once
+
+#include "Core.h"
+
 #include "stb_image.h"
 #include <assimp/Importer.hpp>
 #include <assimp/scene.h>
 #include <assimp/postprocess.h>
 
-#include "TestMesh.h"
-#include "Shader.h"
-
-#include <string>
-#include <fstream>
-#include <sstream>
-#include <iostream>
-#include <map>
-#include <vector>
-using namespace std;
+#include "AssimpMesh.h"
 
 class AssimpModel
 {
 public:
     /*  Model Data */
     vector<Texture> textures_loaded;	// stores all the textures loaded so far, optimization to make sure textures aren't loaded more than once.
-    vector<Mesh> meshes;
+    vector<AssimpMesh> meshes;
     string directory;
     bool gammaCorrection;
     int nodeCount;
     int meshCount;
+
+    /*  Mesh Data  */
+    vector<Vertex> vertices;
+    vector<unsigned int> indices;
+    vector<Texture> textures;
+    unsigned int VAO;
+    glm::mat4 transform;
 
     AssimpModel(const string& path, bool gamma = false);
 
@@ -36,14 +39,11 @@ public:
 
 private:
 
-    void loadModel(string const& path);
+    void initFromScene(string const& path);
     void processAnimNode(aiNodeAnim* node);
     void processNode(aiNode* node, const aiScene* scene);
 
-    Mesh processMesh(aiMesh* mesh, const aiScene* scene);
+    AssimpMesh processMesh(aiMesh* mesh, const aiScene* scene);
     vector<Texture> loadMaterialTextures(aiMaterial* mat, aiTextureType type, string typeName); 
-    unsigned int TextureFromFile(const char* path, const string& directory, bool gamma = false);
+    unsigned int textureFromFile(const char* path, const string& directory, bool gamma = false);
 };
-
-
-#endif
